Rename t/aux -> t/build for Go
Go gets mad about things.  Lots of things.  One of those things is the
use of the path component "aux" on Windows (I'm not on Windows, nor do I
ever intend to be but...).  Sometimes it's easier to give in to the
noisy jerk at the party, than to stand one's ground. =)