Fixed Key Encryption
Vaults (for storing generated encryption ciphers) can now be set to
'fixed key' mode, where they will derive (or use as-is) keys and ivs for
all supported algorithms.  This allows a site to have zero or more
buckets that do _NOT_ randomize encryption keys, in case that sort of
cipher-stability is important.

Note that derived key/iv pairs are still stored in the vault, to
future-proof the encryption against change or rotation of the fixed
keys, and to foster such behavior.

Fixes #12