Support HTTP Parameters in Vault / WebDAV
You can now specify the CA details for connecting to both Vault and
WebDAV (in their respective configuration stanzas) and it will actually
affect how Vault and WebDav are connected to!

Also, we added Timeout support for both, to avoid hard-coding any upper
limits on connection / request / response time frames.  The default is
still "zero timeout", but can now be set to a number of seconds.